feat: code separated into modules
